requirements.txt
pdf2image
pytesseract
fasttext
gensim
pdfminer
slate3k
hachoir
regex
graphviz
unidecode
langdetect
argcomplete==1.10.0
beautifulsoup4==4.8.0
certifi==2019.11.28
chardet==3.0.4
docx2txt==0.8
EbookLib==0.17.1
elasticsearch
extract-msg==0.23.1
ftfy==5.6
hachoir
idna==2.8
IMAPClient==2.1.0
joblib==0.14.1
lxml==4.4.2
nltk==3.4.5
numpy==1.18.1
olefile==0.46
pandas==0.25.3
pdfminer.six==20181108
Pillow==7.0.0
pycryptodome==3.9.4
PyPDF2==1.26.0
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2019.3
requests==2.22.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.12.0
sklearn==0.0
sortedcontainers==2.1.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
textract==1.6.3
tika==1.23
tzlocal==1.5.1
urllib3==1.25.7
wcwidth==0.1.8
xlrd==1.2.0
XlsxWriter==1.2.7
#pip install -e git+git://https://github.com/TakesxiSximada/slate3k